Move SheetContentScrollController to scrollable_sheet_extent.dart
